% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{module_line_dynamic_plot}
\alias{module_line_dynamic_plot}
\title{Plotting a line graph expression dynamic grouped by Modules}
\usage{
module_line_dynamic_plot(cds, module_tbl_complete)
}
\arguments{
\item{cds}{CDS obj generated from Monocle3}

\item{module_tbl_complete}{meta table of genes in each module}
}
\value{
a line graph of class ggplot2
}
\description{
This function loads a CDS obj from Monocle3 and builds a df
that contains cells ordered by pseudotime and calculates the average expression
of genes in each module along pseudotime, returns a line graph
}
